-- temp

+++ src/disk_access_functions.c
@@ -27,7 +27,7 @@ int read_blocks(int db_fd, void* blocks_in_main_memory, uint32_t block_id, uint3
 {
 	off_t start_offset = block_id * block_size;
 	ssize_t bytes_read = pread(db_fd, blocks_in_main_memory, block_count * block_size, start_offset);
-	printf("pread params : %d, %d, %ld, %d\n", db_fd, blocks_in_main_memory, block_count * block_size, start_offset);
+	printf("pread params : %d, %d, %ld, %u\n", db_fd, (int)blocks_in_main_memory, block_count * block_size, start_offset);
 	printf("fd : %d, block_id : %u, blocks_count : %u, block_size, %u, bytes_read : %ld, err : %d\n", db_fd, block_id, block_count, block_size, bytes_read, errno);
 	return bytes_read;
 }
@@ -36,7 +36,7 @@ int write_blocks(int db_fd, void* blocks_in_main_memory, uint32_t block_id, uint
 {
 	off_t start_offset = block_id * block_size;
 	ssize_t bytes_written = pwrite(db_fd, blocks_in_main_memory, block_count * block_size, start_offset);
-	printf("pwrite params : %d, %d, %ld, %d\n", db_fd, blocks_in_main_memory, block_count * block_size, start_offset);
+	printf("pwrite params : %d, %d, %ld, %u\n", db_fd, (int)blocks_in_main_memory, block_count * block_size, start_offset);
 	printf("fd : %d, block_id : %u, blocks_count : %u, block_size, %u, bytes_written : %ld, err : %d\n", db_fd, block_id, block_count, block_size, bytes_written, errno);
 	return bytes_written;
 }
